+++ src/MDVRP/Solution.java
@@ -12,14 +12,14 @@ import java.util.Map;
 public class Solution {
     private List<CrowdedDepot> depots;
     private Individual individual;
-    private Map<Integer, List<Double>> routesDemand;
+    private Map<Integer, List<Integer>> routesDemand;
     private Map<Integer, List<Double>> routesDistance;
 
     public Solution(List<CrowdedDepot> depots, Individual individual, Metrics metrics) {
         this.depots = depots;
         this.individual = individual;
 
-        Map<Integer, List<Double>> routesDemand = new HashMap<>();
+        Map<Integer, List<Integer>> routesDemand = new HashMap<>();
         Map<Integer, List<Double>> routesDistance = new HashMap<>();
 
         for (Map.Entry<Integer, List<Route>> entry : individual.getChromosome().entrySet()) {
@@ -28,7 +28,7 @@ public class Solution {
             routesDistance.put(depotID, new ArrayList<>());
             List<Route> chromosomeDepot = entry.getValue();
             for (Route route: chromosomeDepot) {
-                double routeDemand = metrics.getRouteDemand(route.getRoute());
+                int routeDemand = metrics.getRouteDemand(route.getRoute());
                 double routeDistance = metrics.getRouteDistance(depotID, route.getRoute());
 
                 routesDemand.get(depotID).add(routeDemand);
@@ -42,6 +42,6 @@ public class Solution {
     // Getters
     public Individual getIndividual()                       { return individual; }
     public List<CrowdedDepot> getDepots()                   { return depots; }
-    public Map<Integer, List<Double>> getRoutesDemand()     { return routesDemand; }
+    public Map<Integer, List<Integer>> getRoutesDemand()     { return routesDemand; }
     public Map<Integer, List<Double>> getRoutesDistance()   { return routesDistance; }
 }
